src/kitchenai_community/llama_index_toolhouse/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
apturl==0.5.2
argcomplete==2.1.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
asyncio==3.4.3
attrs==24.2.0
Authlib==1.3.1
autoflake==2.3.1
Babel==2.13.1
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
binaryornot==0.4.4
black==24.8.0
bleach==6.1.0
blinker==1.4
boto3==1.35.49
botocore==1.35.49
Brlapi==0.8.3
build==1.2.2.post1
cachetools==5.5.0
cappa==0.22.5
certifi==2024.8.30
cffi==1.16.0
cfgv==3.3.1
chardet==4.0.0
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.5.15
click==8.1.7
colorama==0.4.4
coloredlogs==15.0.1
comm==0.2.0
command-not-found==0.3
cookiecutter==2.6.0
cruft==2.15.0
cryptography==3.4.8
cssbeautifier==1.15.1
cupshelpers==1.0
dataclasses-json==0.6.7
dbus-python==1.2.18
ddgr==1.9
debugpy==1.8.0
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.14
dirtyjson==1.0.8
distlib==0.3.8
distro==1.9.0
distro-info==1.1+ubuntu0.2
Django==5.1.2
django-ninja==1.3.0
djlint==1.35.2
duplicity==0.8.21
durationpy==0.9
EditorConfig==0.12.4
evdev==1.4.0
exceptiongroup==1.2.0
executing==2.0.1
falco-cli==0.26.6
fastapi==0.115.3
fasteners==0.14.1
fastjsonschema==2.19.0
filelock==3.16.1
flatbuffers==24.3.25
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.10.0
future==0.18.2
gitdb==4.0.11
GitPython==3.1.43
-e git+https://github.com/weaviate/Verba@a11c0d0f756e489253e300873d0563234ad413c9#egg=goldenverba
google-auth==2.35.0
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.67.0
h11==0.14.0
hatch==1.12.0
hatch-pip-compile==1.11.3
hatchling==1.25.0
hidpidaemon==18.4.6
honcho==1.1.0
html-tag-names==0.1.2
html-void-elements==0.1.0
httpcore==1.0.6
httplib2==0.20.2
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.1
humanfriendly==10.0
hyperlink==21.0.0
identify==2.4.10
idna==3.10
importlib_metadata==8.4.0
importlib_resources==6.4.5
ipykernel==6.26.0
ipython==8.18.0
isoduration==20.11.0
isort==5.13.2
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.2
jiter==0.6.1
jmespath==1.0.1
joblib==1.4.2
jsbeautifier==1.15.1
json-dotenv==0.0.29
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter-events==0.9.0
jupyter-lsp==2.2.0
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.10.1
jupyter_server_terminals==0.4.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
just-bin==1.26.0
keyring==23.5.0
kubernetes==31.0.0
langchain-community==0.0.24
langchain-core==0.1.28
langsmith==0.1.11
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
llama-cloud==0.1.4
llama-index==0.11.20
llama-index-agent-openai==0.3.4
llama-index-cli==0.3.1
llama-index-core==0.11.20
llama-index-embeddings-openai==0.2.5
llama-index-indices-managed-llama-cloud==0.4.0
llama-index-legacy==0.9.48.post3
llama-index-llms-openai==0.2.16
llama-index-multi-modal-llms-openai==0.2.3
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.2
llama-index-readers-llama-parse==0.3.0
llama-index-vector-stores-chroma==0.2.1
llama-parse==0.5.12
lockfile==0.12.2
louis==3.20.0
lxml==5.2.2
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==3.0.0
MarkupSafe==2.0.1
marshmallow==3.23.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mmh3==5.0.1
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.0.3
multidict==6.1.0
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.4.2
nltk==3.9.1
nodeenv==0.13.4
notebook_shim==0.2.3
numpy==1.26.4
oauthlib==3.2.2
olefile==0.46
onnxruntime==1.19.2
openai==1.52.2
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.10
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
pathspec==0.12.1
pexpect==4.8.0
pillow==11.0.0
pip-tools==7.4.1
pipx==1.2.0
platformdirs==4.0.0
pluggy==1.5.0
posthog==3.7.0
pre-commit==2.17.0
prometheus-client==0.19.0
prompt-toolkit==3.0.41
propcache==0.2.0
protobuf==4.25.5
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydantic==2.9.2
pydantic_core==2.23.4
pydbus==0.6.0
pyflakes==3.2.0
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pynvim==0.4.2
pyparsing==2.4.7
pypdf==4.3.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-docx==1.1.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-slugify==8.0.4
python-xlib==0.29
pytz==2024.2
pyxdg==0.27
PyYAML==6.0.2
pyzmq==25.1.1
referencing==0.31.0
regex==2024.9.11
reportlab==3.6.8
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.3
rpds-py==0.13.1
rsa==4.9
s3transfer==0.10.3
SecretStorage==3.3.1
Send2Trash==1.8.2
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
speedtest-cli==2.1.3
SQLAlchemy==2.0.36
sqlparse==0.5.1
stack-data==0.6.3
starlette==0.41.2
striprtf==0.0.26
sympy==1.13.3
system76driver==20.4.95
systemd-python==234
tenacity==8.5.0
terminado==0.18.0
text-unidecode==1.3
tiktoken==0.8.0
tinycss2==1.2.1
tokenizers==0.20.1
toml==0.10.2
tomli==2.0.1
tomli_w==1.0.0
tomlkit==0.13.2
toot==0.36.0
tornado==6.3.3
tqdm==4.66.5
traitlets==5.13.0
trove-classifiers==2024.9.12
typer==0.12.5
types-python-dateutil==2.8.19.14
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uri-template==1.3.0
urllib3==2.2.3
urwid==2.1.2
usb-creator==0.3.7
userpath==1.8.0
uv==0.4.15
uvicorn==0.32.0
uvloop==0.21.0
validators==0.21.0
vboxapi==1.0
virtualenv==20.26.5
visidata==2.2.1
wadllib==1.3.6
wasabi==1.1.2
watchfiles==0.24.0
wcwidth==0.2.6
weaviate-client==3.23.1
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
wrapt==1.16.0
xdg==5
xkit==0.0.0
yarl==1.16.0
zipp==3.20.2
zstandard==0.23.0
